improving exceptions about readme.rst
now, we're requiring `README.rst` when running 'require', 'bdist_wheel'
and `sdist`. the last two is necessary because `twine` uses these dists
to get the metadata to submit (register) to PyPI.

it was also created a workaround for `check-manifest`. this was
necessary because `check-manifest` generates a temp `sdist` to your
checks.